delete highlevel Card componets and render Card from Event
