+++ src/components/events/ImageEvent.js
@@ -1,8 +1,29 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import styled from 'styled-components';
 import Event from '../events/Event';
-import ImageCard from '../cards/ImageCard';
+import ConditionalMarkdown from '../atoms/ConditionalMarkdown';
+import ImageAtom from '../atoms/ImageAtom';
 import { joinClassNames } from '../../utils/classNames';
+
+const ImageText = styled(ConditionalMarkdown)(props => {
+  const defaults = {};
+
+  const style = { ...defaults, ...props.theme.imageText };
+
+  return style;
+});
+
+const ImageCredit = styled(ConditionalMarkdown)(props => {
+  const defaults = {
+    marginTop: '10px',
+    fontSize: '0.85rem',
+  };
+
+  const style = { ...defaults, ...props.theme.imageCredit };
+
+  return style;
+});
 
 export default function ImageEvent(props) {
   const { date, src, alt, credit, text, marker, children, className } = props;
@@ -13,9 +34,10 @@ export default function ImageEvent(props) {
       date={date}
       marker={marker}
     >
-      <ImageCard date={date} src={src} alt={alt} credit={credit} text={text}>
-        {children}
-      </ImageCard>
+      <ImageText className="image-text">{text}</ImageText>
+      <ImageAtom src={src} alt={alt} />
+      {credit && <ImageCredit className="image-credit">{credit}</ImageCredit>}
+      {children}
     </Event>
   );
 }
