+++ src/components/events/YouTubeEvent.js
@@ -1,8 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import styled from 'styled-components';
 import Event from './Event';
-import YouTubeCard from '../cards/YouTubeCard';
+import ConditionalMarkdown from '../atoms/ConditionalMarkdown';
+import YouTubeAtom from '../atoms/YouTubeAtom';
 import { joinClassNames } from '../../utils/classNames';
+
+const YouTubeText = styled(ConditionalMarkdown)(props => {
+  const defaults = {};
+
+  const style = { ...defaults, ...props.theme.youTubeText };
+
+  return style;
+});
 
 export default function YouTubeEvent(props) {
   const { date, id, name, text, marker, children, className } = props;
@@ -13,9 +23,9 @@ export default function YouTubeEvent(props) {
       date={date}
       marker={marker}
     >
-      <YouTubeCard date={date} id={id} name={name} text={text}>
-        {children}
-      </YouTubeCard>
+      <YouTubeText className="youtube-text">{text}</YouTubeText>
+      <YouTubeAtom id={id} name={name} />
+      {children}
     </Event>
   );
 }
