--- src/hooks/useIntersectionObserver.js
@@ -1,22 +1,0 @@
-import { useEffect, useRef, useState } from 'react';
-import 'intersection-observer';
-
-export default function useIntersectionObserver(opts) {
-  const [isVisible, setIsVisible] = useState(false);
-  const ref = useRef(null);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(entries => {
-      entries.forEach(entry => {
-        if (entry.isIntersecting) {
-          setIsVisible(true);
-        } else {
-          setIsVisible(false);
-        }
-      });
-    }, opts);
-    observer.observe(ref.current);
-  });
-
-  return [isVisible, ref];
-}
