+++ src/index.js
@@ -2,20 +2,14 @@ import Button from './components/Button';
 import EventMarker from './components/EventMarker';
 import Events from './components/Events';
 import Timeline from './components/Timeline';
-import useIntersectionObserver from './hooks/useIntersectionObserver';
 import themes from './themes';
 
 import TextAtom from './components/atoms/TextAtom';
 import TweetAtom from './components/atoms/TweetAtom';
 import YouTubeAtom from './components/atoms/YouTubeAtom';
 
-import Card from './components/cards/Card';
-import ImageCard from './components/cards/ImageCard';
-import TextCard from './components/cards/TextCard';
-import TweetCard from './components/cards/TweetCard';
-import YouTubeCard from './components/cards/YouTubeCard';
-
 import Event from './components/events/Event';
+import Card from './components/events/Card';
 import ImageEvent from './components/events/ImageEvent';
 import TextEvent from './components/events/TextEvent';
 import TweetEvent from './components/events/TweetEvent';
@@ -30,16 +24,11 @@ export {
   Events,
   Card,
   ImageEvent,
-  ImageCard,
   TextEvent,
   TextAtom,
-  TextCard,
   TweetAtom,
-  TweetCard,
   TweetEvent,
   YouTubeAtom,
-  YouTubeCard,
   YouTubeEvent,
-  useIntersectionObserver,
   themes,
 };
